+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unrelease]
+- 0307: Fixed issue with Details page Image component's fallback is not used for non-image assets. 
+- 0308: Fixed issue where unsupported (by the browser) image asset types (ex. DFX) are used for image display in browser (thumbs/preview) instead of placeholder.
+
 ## [v1.6.9]
 
 ### Fixed
--0301: Files having special character in the filename. Download,Share and Add to Cart do not work.
+- 0301: Files having special character in the filename. Download,Share and Add to Cart do not work.
 
 ## [Unreleased]